Operaciones_fichero

+++ Capítulo_10/operaciones_fichero.py
@@ -1,4 +1,5 @@
 import os
+import pathlib
 import sys
 sys.path.append('../modulos/')
 from modulos.file import File
@@ -7,19 +8,20 @@ from modulos.switch_flag import set_flag_es
 def menu_principal():
     respuesta = 0
     #Función para tratar el menú del programa
-    while respuesta not in (1, 2, 3):
+    while respuesta not in (1, 2, 3, 4, 5):
         os.system('clear')
         print("---------- Operaciones con ficheros .txt -----------\n")
         print("1. Leer contenido del fichero.")
         print("2. Escribir en el fichero.")
         print("3. Modificar o borrar fichero.")
-        print("4. Salir del programa.")
+        print("4. Listar ficheros.")
+        print("5. Salir del programa.")
         try:
             #Tratamos el error por si meten alguna opción que haga romper
             respuesta = int(input("\nOpción: "))
         except ValueError:
             respuesta = 0
-        if respuesta in (1, 2, 3, 4):
+        if respuesta in (1, 2, 3, 4, 5):
             return respuesta
         else:
             print("No es una opción válida.")
@@ -30,16 +32,21 @@ def menu_lectura():
     #Función para mostrar el menú de lectura y leer el fichero
     os.system('clear')
     file = input('Introduce el nombre del fichero que quieres leer: ')
-    fichero = File(file)
+    fichero = File(file.lower())
     os.system('clear')
-    fichero.leer_fichero()
+    if fichero.existe():
+        fichero.leer_fichero()
 
 def menu_escritura():
     os.system('clear')
+    file = ''
     #Función para mostrar el menú de escritura y escribir en el fichero
-    file = input('Introduce el nombre del fichero en el que deseas escribir: ')
+    print("En caso de no existir, se creará el fichero.\n")
+    while file == '':
+        file = input(
+            'Introduce el nombre del fichero en el que deseas escribir: ')
     cadena = input('Introduce el texto a escribir en el fichero: ')
-    fichero = File(file)
+    fichero = File(file.lower())
     fichero.escribir_fichero(cadena)
 
 def menu_borrado_fichero():
@@ -57,25 +64,38 @@ def menu_borrado_fichero():
     if opcion == 1:
         os.system('clear')
         file = input('Introduce el nombre del fichero que quieres modificar: ')
-        cadena = input('Introduce la frase o palabra que quieras modificar: ')
-        cadena_new = input('Introduce la frase por la que quieres cambiarla: ')
-        fichero = File(file)
-        fichero.sustituir_cadena(cadena, cadena_new)
+        fichero = File(file.lower())
+        if fichero.existe():
+            cadena = input('Introduce la frase o palabra que quieras modificar: ')
+            cadena_new = input('Introduce la frase por la que quieres cambiarla: ')
+            fichero.sustituir_cadena(cadena, cadena_new)
     elif opcion == 2:
         os.system('clear')
         file = input('Introduce el nombre del fichero que quieres borrar: ')
-        fichero = File(file)
-        fichero.sustituir_cadena(fichero.contenido)        
+        fichero = File(file.lower())
+        if fichero.existe():
+            fichero.sustituir_cadena(fichero.contenido)        
     elif opcion == 3:
         os.system('clear')
         file = input('Introduce el nombre del fichero que quieres eliminar: ')
-        fichero = File(file)
-        fichero.eliminar_fichero()
+        fichero = File(file.lower())
+        if fichero.existe():
+            fichero.eliminar_fichero()
+
+def menu_listar():
+    #Función para listar los archivos txt del directorio con el que trabajamos
+    os.system('clear')
+    p = pathlib.Path('/Users/agustined/Desktop/python_work/files/')
+    print("------------- Ficheros en el directorio -------------------\n")
+    for file in p.glob('*.txt'):
+        file = f"{file}"
+        print(file.removeprefix('/Users/agustined/Desktop/python_work/files/'))
+    input("\nPulse enter para salir...")
         
 
 #------------------ Programa principal ---------------------------------------
 opcion = 0
-while opcion != 4:
+while opcion != 5:
     opcion = menu_principal()
     if opcion == 1:
         menu_lectura()
@@ -83,4 +103,6 @@ while opcion != 4:
         menu_escritura()
     elif opcion == 3:
         menu_borrado_fichero()
+    elif opcion == 4:
+        menu_listar()
 os.system('clear')
